// ui/src/generated/models/policyAction.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Lintel
 * OpenAPI spec version: 0.1.0
 */

export type PolicyAction = typeof PolicyAction[keyof typeof PolicyAction];


export const PolicyAction = {
  require_approval: 'require_approval',
  auto_approve: 'auto_approve',
  block: 'block',
  notify: 'notify',
} as const;